Move types to `RoomToolboxContext`
