+++ apps/meteor/client/hooks/roomActions/useAutotranslateRoomAction.ts
@@ -1,7 +1,7 @@
 import { useSetting, usePermission } from '@rocket.chat/ui-contexts';
 import { lazy, useMemo } from 'react';
 
-import type { ToolboxActionConfig } from '../../views/room/lib/Toolbox';
+import type { ToolboxActionConfig } from '../../views/room/contexts/RoomToolboxContext';
 
 const AutoTranslate = lazy(() => import('../../views/room/contextualBar/AutoTranslate'));
