+++ apps/meteor/client/hooks/roomActions/useKeyboardShortcutListRoomAction.ts
@@ -1,6 +1,6 @@
 import { lazy, useMemo } from 'react';
 
-import type { ToolboxActionConfig } from '../../views/room/lib/Toolbox';
+import type { ToolboxActionConfig } from '../../views/room/contexts/RoomToolboxContext';
 
 const KeyboardShortcuts = lazy(() => import('../../views/room/contextualBar/KeyboardShortcuts'));
